Improve c++ code
