+++ deeppanel/src/main/cpp/Panel.hpp
@@ -1,0 +1,23 @@
+//
+//  Panel.hpp
+//  DeepPanel
+//
+//  Created by Pedro Gómez on 11/10/2020.
+//  Copyright © 2020 Pedro Gómez. All rights reserved.
+//
+
+#ifndef Panel_hpp
+#define Panel_hpp
+
+#include <stdio.h>
+
+class Panel {
+public:
+    int left;
+    int bottom;
+    int right;
+    int top;
+};
+
+
+#endif /* Panel_hpp */
