+++ deeppanel/src/main/cpp/connected-components.hpp
@@ -1,0 +1,7 @@
+#ifndef connected_components_h
+#define connected_components_h
+#import "ConnectedComponentResult.hpp"
+
+ConnectedComponentResult find_components(int **matrix, int m, int n);
+
+#endif /* connected_components_h */
